Add slight delay to Agent movement
The player has a slight delay when in block-movement mode; because when
in arrow-movement mode the delay is handled by the arrows, we don't
impose an additional delay.

Because the agent was derived from player code, it was erroneously
giving itself no delay when the _player_ was in arrow mode.

Re-added the delay, independent of the state. Also gave it a delay of
about 2/3 that of the player to represent the fact that the agent should
still be moving slightly faster