Add "power rails" step to redstone updates
Utilizing the new concept of Adjacency Sets, we introduce a custom
definition of powered rail adjacency that only considered rails that are
connected, and divide all the powerable rails on the plane up into
"connectivity sets". We then simply add power to every rail in a
connectivity set that has any rail receiving power.